requirements.txt
python-dotenv
langchain
langchain-community
langchain-groq
faiss-cpu
tqdm
pandas
unstructured
nltk
pillow
PyMuPDF
beautifulsoup4
scikit-learn

# Voice Recognition Dependencies
SpeechRecognition==3.10.0
pydub==0.25.1
pyttsx3==2.90
gTTS==2.3.2
langdetect==1.0.9
deep-translator

# OCR Dependencies  
pytesseract
opencv-python

# Telegram Bot
python-telegram-bot==20.7